requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
openai>=1.0.0
azure-search-documents>=11.4.0
azure-identity>=1.15.0
python-dotenv>=1.0.0